limit to user

+++ datastore/blogs.js
@@ -12,7 +12,7 @@ exports.readBlogList = function (req,res,next) {
 
   console.log("blogList read");
   var blogs = req.db.collection("blogs");
-  blogs.find(function (error,result) {
+  blogs.find({userID:req.session.userID},function (error,result) {
     if (error) {
       next(error);
       return;
@@ -78,7 +78,7 @@ exports.deleteBlog = function (req,res,next) {
 
     var blogs = req.db.collection("blogs");
     console.log("obj",obj);
-    blogs.deleteOne({_id:new ObjectID(obj._id)}, function (err, res2) {
+    blogs.deleteOne({_id:new ObjectID(obj._id),userID:req.session.userID}, function (err, res2) {
       if (err) {
         return next(err);
       }
@@ -110,6 +110,7 @@ exports.readBlog = function (req,res,next) {
     if (obj._id) {
       obj._id = new ObjectID(obj._id);
     }
+    obj.userID = req.session.userID;
     console.log("criterion",obj);
     blogs.findOne(obj, function (err,res2) {
       if (err) {
@@ -126,7 +127,7 @@ exports.readBlog = function (req,res,next) {
       console.log("blogID",blog._id);
       console.log("blogID",blog._id.toString());
       console.log("blogID type",typeof blog._id);
-      posts.find({blogID:blog._id.toString()}, function (err,res3) {
+      posts.find({blogID:blog._id.toString(),userID:req.session.userID}, function (err,res3) {
         res3.toArray(function (err, list) {
           console.log("toArray err",err);
           console.log("toArray list",list);
@@ -160,7 +161,7 @@ exports.saveBlogTitle = function (req,res,next) {
       title: obj.title,
       userID: req.session.userID
     };
-    blogs.updateOne({_id: new ObjectID(obj._id)}, blogObj, function (err, res2) {
+    blogs.updateOne({_id: new ObjectID(obj._id),userID:req.session.userID}, blogObj, function (err, res2) {
       if (err) {
         return next(err);
       }
