+++ datastore/posts.js
@@ -18,7 +18,7 @@ exports.updatePost = function (req,res,next) {
     post.loadObject(obj);
     var postObj = post.exportObject(true, true);
     postObj.userID = req.session.userID;
-    posts.updateOne({_id:new ObjectID(post._id)}, postObj, function (err,res2) {
+    posts.updateOne({_id:new ObjectID(post._id),userID:req.session.userID}, postObj, function (err,res2) {
       if (err) {
         return next(err);
       }
@@ -69,7 +69,7 @@ exports.deletePost = function (req,res,next) {
     }
 
     var posts = req.db.collection("posts");
-    posts.deleteOne({_id:new ObjectID(obj._id)}, function (err,res2) {
+    posts.deleteOne({_id:new ObjectID(obj._id),userID:req.session.userID}, function (err,res2) {
       if (err) {
         return next(err);
       }
